Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main{
-    int a;
+    float a;
     printf("Enter float no.");
     scanf("%f",&a);
-    printf("%f",a); 
+    printf("%.2f",a); 
     return 0;
 }
